+++ src/kit.ts
@@ -716,6 +716,11 @@ export async function getShellScriptEnvironment(kit: Kit, opts?: expand.Expansio
     await fs.unlink(script_path); // delete script file
     const output = (res.stdout) ? res.stdout + (res.stderr || '') : res.stderr;
 
+    if (res.retc !== 0) {
+        log.error(localize('error.running.setup.script', 'Error running {0} with: {1}', kit.environmentSetupScript, output));
+        return;
+    }
+
     let env = '';
     try {
         /* When the script failed, envpath would not exist */